services/plateaubreaker/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
pika==1.3.2
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.1 